refactor: display ip as is if it is docker-ip and extract into HeaderPagecontroller class
